Adding np and ak setops. Renaming file
