+++ benchmark.ini
@@ -15,7 +15,7 @@ testpaths =
     benchmark_v2/encoding_benchmark.py
     benchmark_v2/gather_benchmark.py
     benchmark_v2/scatter_benchmark.py
-    benchmark_v2/segarr_setops_benchmark.py
+    benchmark_v2/setops_benchmark.py
 python_functions = bench_*
 env =
     D:ARKOUDA_SERVER_HOST=localhost
